admin sessions

+++ vymazat-reprizu.php
@@ -1,9 +1,16 @@
 <?php
+session_start();
+if (!isset($_SESSION['admin_logged_in']) || $_SESSION['admin_logged_in'] !== true) {
+    header('Location: index.php'); // alebo prihlasenie.php
+    exit;
+}
 require_once 'db/config.php';
 require_once 'classes/Database.php';
 
 $database = new \Classes\Database();
 $conn = $database->getConnection();
+
+
 
 if (isset($_GET['id'])) {
     $stmt = $conn->prepare("DELETE FROM reprizy WHERE id = :id");
